+++ modules/ar_system.c
@@ -219,7 +219,7 @@ bool ar_system__process_next_message(ar_system_t *mut_system) {
     // Take ownership of the message for the system
     ar_data__take_ownership(own_message, mut_system);
 
-    ar_interpreter__execute_method(mut_system->own_interpreter, agent_id, own_message);
+    ar_interpreter__execute_method(mut_system->own_interpreter, agent_id, own_message, mut_system);
 
     if (ar_agency__agent_has_messages(mut_system->own_agency, agent_id)) {
         mut_system->next_agent_hint = agent_id;